Move findInactiveWorkplaces to new file
