+++ server/routes/admin/email-campaigns/inactive-workplaces/findInactiveWorkplaces.js
@@ -1,0 +1,28 @@
+const findInactiveWorkplaces = async () => {
+  return [
+    {
+      name: 'Workplace Name',
+      nmdsId: 'J1234567',
+      lastUpdated: '2020-06-01',
+      emailTemplate: 6,
+      dataOwner: 'Workplace',
+      user: {
+        name: 'Test Name',
+        email: 'test@example.com',
+      },
+    },
+    {
+      name: 'Second Workplace Name',
+      nmdsId: 'A0012345',
+      lastUpdated: '2020-01-01',
+      emailTemplate: 12,
+      dataOwner: 'Workplace',
+      user: {
+        name: 'Name McName',
+        email: 'name@mcname.com',
+      },
+    }
+  ];
+}
+
+module.exports = findInactiveWorkplaces;
