+++ server/test/unit/routes/admin/email-campaigns/inactive-workplaces/index.spec.js
@@ -1,8 +1,7 @@
 const expect = require('chai').expect;
-const { findInactiveWorkplaces } = require('../../../../../../routes/admin/email-campaigns/inactive-workplaces');
+const findInactiveWorkplaces = require('../../../../../../routes/admin/email-campaigns/inactive-workplaces/findInactiveWorkplaces');
 
-describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
-
+describe.only('server/routes/admin/email-campaigns/inactive-workplaces', () => {
   it('should return inactive workplaces', async () => {
     const inactiveWorkplaces = await findInactiveWorkplaces();
